+++ src/src/transports/smtp.h
@@ -105,7 +105,7 @@ extern void smtp_transport_closedown(transport_instance *);
 extern int     smtp_auth(uschar *, unsigned, address_item *, host_item *,
 		 smtp_transport_options_block *, BOOL,
 		 smtp_inblock *, smtp_outblock *);
-extern BOOL    smtp_mail_auth_str(uschar *, unsigned,
+extern BOOL    smtp_mail_auth_str(uschar *, size_t,
 		 address_item *, smtp_transport_options_block *);
 
 #ifdef SUPPORT_SOCKS
